requirements.txt
pandas
numpy
flask
mysql-connector-python
openpyxl
pywin32
chardet
werkzeug